+++ client/src/lib/state/power-vs/power-vs.js
@@ -65,6 +65,11 @@ function powerVsOnStoreUpdate(config) {
     workspace.images = selectedImages;
     // add unfound networks to attachments
     workspace.network.forEach((nw) => {
+      if (nw.pi_network_jumbo && !nw.pi_network_mtu) {
+        nw.pi_network_mtu = "9000";
+      } else if (!nw.pi_network_mtu) {
+        nw.pi_network_mtu = "";
+      }
       if (!splatContains(workspace.attachments, "network", nw.name)) {
         workspace.attachments.push({
           network: nw.name,
@@ -313,7 +318,7 @@ function initPowerVsStore(store) {
         delete: powerVsNetworkDelete,
         save: powerVsNetworkSave,
         shouldDisableSave: shouldDisableComponentSave(
-          ["name", "pi_cidr", "pi_dns", "pi_network_type"],
+          ["name", "pi_cidr", "pi_dns", "pi_network_type", "pi_network_mtu"],
           "power",
           "network"
         ),